refactor(Vsix): match namespace to "Shared" folder.
