+++ src/LibraryManager.Vsix/Commands/RestoreOnBuildCommand.cs
@@ -11,6 +11,7 @@ using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.Vsix.Contracts;
+using Microsoft.Web.LibraryManager.Vsix.Shared;
 using NuGet.VisualStudio;
 using Task = System.Threading.Tasks.Task;
