+++ src/LibraryManager.Vsix/Json/SuggestedActions/SuggestedActionProvider.cs
@@ -9,6 +9,7 @@ using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Utilities;
 using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.Vsix.Contracts;
+using Microsoft.Web.LibraryManager.Vsix.Shared;
 using Microsoft.WebTools.Languages.Json.Parser.Nodes;
 using Microsoft.WebTools.Languages.Json.VS.SuggestedActions;
 using Microsoft.WebTools.Languages.Shared.Parser.Nodes;
