+++ src/LibraryManager.Vsix/Json/SuggestedActions/UpdateSuggestedAction.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using Microsoft.VisualStudio.Imaging;
 using Microsoft.VisualStudio.Text;
 using Microsoft.Web.LibraryManager.Contracts;
+using Microsoft.Web.LibraryManager.Vsix.Shared;
 using Microsoft.WebTools.Languages.Json.Parser.Nodes;
 using Microsoft.WebTools.Languages.Shared.Editor.SuggestedActions;
 using Microsoft.WebTools.Languages.Shared.Parser.Nodes;
