+++ src/LibraryManager.Vsix/Shared/VsHelpers.cs
@@ -19,7 +19,7 @@ using Microsoft.VisualStudio.Telemetry;
 using Microsoft.Web.LibraryManager.Contracts;
 using Task = System.Threading.Tasks.Task;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Shared
 {
     internal static class VsHelpers
     {
