+++ src/LibraryManager.Vsix/Shared/WpfUtil.cs
@@ -14,7 +14,7 @@ using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using Color = System.Windows.Media.Color;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Shared
 {
     internal static class WpfUtil
     {
